+++ docs/release_notes.rst
@@ -9,8 +9,9 @@ v2025.12.x (2025-12-XX)
 New Data
 ^^^^^^^^
 
- * contains cleaned company-wide natural gas disposition data from Part 6B
-   of the EIA 176 survey. See :pr:`4765`
+ * Added :ref:`core_eia176__yearly_gas_disposition`, which contains cleaned
+   company-wide natural gas disposition data from Part 6B of the EIA 176 survey. See
+   :issue:`4708` and :pr:`4765`. Thanks to :user:`MeadBarrel`!
 
 
 Expanded Data Coverage